requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
playwright>=1.40.0
httpx>=0.27.0
pyttsx3>=2.90
sounddevice>=0.4.7
vosk>=0.3.45
numpy>=1.26.0
